fix(upgrade): 普通用户升级兼容
