+++ lib/cluster.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 import json
 import re
-
+from getpass import getuser
 from .ssh import SSHClient
 from . import k8s
 
@@ -32,7 +32,7 @@ class OnecloudCluster(object):
     def __init__(self, ssh_client):
         self.ssh_client = ssh_client
         ret = ssh_client.exec_command(
-            'kubectl -n onecloud get onecloudclusters default -o json')
+            'sudo kubectl -n onecloud get onecloudclusters default -o json')
         try:
             cluster = json.loads(ret)
         except ValueError:
@@ -85,7 +85,7 @@ class OnecloudCluster(object):
         return self.get_spec().get('version')
 
     def _construct_nodes(self):
-        k8s_nodes = json.loads(self.ssh_client.exec_command('kubectl get nodes -o json')).get('items')
+        k8s_nodes = json.loads(self.ssh_client.exec_command('sudo kubectl get nodes -o json')).get('items')
         self.k8s_nodes = [k8s.Node(obj) for obj in k8s_nodes]
         self.master_nodes = [node for node in self.k8s_nodes if node.is_master()]
         self.worker_nodes = [node for node in self.k8s_nodes if not node.is_master()]
@@ -123,7 +123,7 @@ class OnecloudCluster(object):
         return inventory.generate_content()
 
     def set_current_version(self, version):
-        cmd = 'kubectl -n onecloud annotate --overwrite=true onecloudclusters default %s=%s' % (
+        cmd = 'sudo kubectl -n onecloud annotate --overwrite=true onecloudclusters default %s=%s' % (
             A_OCBOOT_UPGRADE_CURRENT_VERSION,
             version)
         self.ssh_client.exec_command(cmd)
@@ -178,7 +178,7 @@ class AnsibleInventory(object):
 
 class ansibleHost(object):
 
-    def __init__(self, node, role, user='root', port=22):
+    def __init__(self, node, role, user=getuser(), port=22):
         self.hostname = node.get_hostname()
         self.ip = node.get_ip()
         self.role = role
@@ -215,20 +215,20 @@ class ansibleHost(object):
 
 class AnsiblePrimaryMasterHost(ansibleHost):
 
-    def __init__(self, node, user='root', port=22):
+    def __init__(self, node, user=getuser(), port=22):
         super(AnsiblePrimaryMasterHost, self).__init__(
             node, GROUP_PRIMARY_MASTER_NODE, user, port)
 
 
 class AnsibleMasterHost(ansibleHost):
 
-    def __init__(self, node, user='root', port=22):
+    def __init__(self, node, user=getuser(), port=22):
         super(AnsibleMasterHost, self).__init__(
             node, GROUP_MASTER_NODES, user, port)
 
 
 class AnsibleWorkerHost(ansibleHost):
 
-    def __init__(self, node, user='root', port=22):
+    def __init__(self, node, user=getuser(), port=22):
         super(AnsibleWorkerHost, self).__init__(
             node, GROUP_WORKER_NODES, user, port)
